Struct kurinji::Kurinji

pub struct Kurinji { /* fields omitted */ }

Resource to access all Kurinji APIs

Implementations

impl Kurinji

pub fn get_bindings_from_json_file(path: &str) -> Bindings

pub fn get_bindings_from_ron_file(path: &str) -> Bindings

pub fn get_bindings_from_json(json: &str) -> Bindings

pub fn get_bindings_from_ron(ron: &str) -> Bindings

impl Kurinji

pub fn push(&mut self, bindings: Bindings)

Push bindings to stack

pub fn push_additive(&mut self, bindings: Bindings)

Clones current bindings and pushes it with the new changes

pub fn pop(&mut self)

Pop the current binding from stack

impl Kurinji

pub fn get_bindings(&self) -> Bindings

pub fn set_bindings(&mut self, bindings: Bindings)

impl Kurinji

pub const MAX_PLAYER_HANDLES: usize

Max number of players that can be connected using gamepads

impl Kurinji

pub fn get_event_phase(&self, action: &str) -> &EventPhase

Returns in which event phase this action active will be true

pub fn set_event_phase(
&mut self,
action: &str,
phase: EventPhase
) -> &mut Kurinji

Set on which event phase should action will be true. By default will be Phase::OnProgress

impl Kurinji

pub fn get_action_strength(&self, action: &str) -> f32

Provides strength of action in range 0.0 - 1.0. Useful when working actions that mapped to analog input eg. joystick

pub fn is_action_active(&self, action: &str) -> bool

Is this action happening. Note* this depends on action event phase

pub fn set_dead_zone(&mut self, action: &str, value: f32) -> &mut Kurinji

Set a dead zone threshold i.e. strenght will be 0.0 until threshold is met. The strength range 0.0 - 1.0 is now mapped to min_threshold - 1.0

Note* meaningful only for analog inputs like joystick, mouse move delta...etc

pub fn set_strength_curve_function(
&mut self,
action: &str,
function: fn(_: f32) -> f32
) -> &mut Kurinji

Set a custom curve function that will be applied to actions strength.

impl Kurinji

pub fn bind_gamepad_button_pressed(
&mut self,
pad_button: GamepadButtonType,
action: &str
) -> &mut Kurinji

pub fn bind_gamepad_button_pressed_for_player(
&mut self,
player: usize,
pad_button: GamepadButtonType,
action: &str
) -> &mut Kurinji

pub fn unbind_gamepad_button_pressed(
&mut self,
pad_button: GamepadButtonType
) -> &mut Kurinji

pub fn unbind_gamepad_button_pressed_for_player(
&mut self,
player: usize,
pad_button: GamepadButtonType
) -> &mut Kurinji

pub fn bind_gamepad_axis(
&mut self,
axis: GamepadAxis,
action: &str
) -> &mut Kurinji

pub fn bind_gamepad_axis_for_player(
&mut self,
player: usize,
axis: GamepadAxis,
action: &str
) -> &mut Kurinji

pub fn unbind_gamepad_axis(&mut self, pad_axis: GamepadAxis) -> &mut Kurinji

pub fn unbind_gamepad_axis_for_player(
&mut self,
player: usize,
axis: GamepadAxis
) -> &mut Kurinji

impl Kurinji

pub fn bind_keyboard_pressed(
&mut self,
code: KeyCode,
action: &str
) -> &mut Kurinji

pub fn unbind_keyboard_pressed(&mut self, code: KeyCode) -> &mut Kurinji

impl Kurinji

pub fn bind_mouse_button_pressed(
&mut self,
code: MouseButton,
action: &str
) -> &mut Kurinji

pub fn unbind_mouse_button_pressed(
&mut self,
button: MouseButton
) -> &mut Kurinji

pub fn bind_mouse_motion(
&mut self,
axis: MouseAxis,
action: &str
) -> &mut Kurinji

pub fn unbind_mouse_motion(&mut self, axis: MouseAxis) -> &mut Kurinji

impl Kurinji

pub fn get_bindings_as_ron(&self) -> Result<String, String>

pub fn set_bindings_with_ron(&mut self, ron: &str) -> &mut Kurinji

pub fn get_bindings_as_json(&self) -> Result<String, String>

pub fn set_bindings_with_json(&mut self, json: &str) -> &mut Kurinji